Moved symbology items out of loader
